src/test/java/com/luxoft/shop/dao/jdbc/JdbcProductDaoTest.java
package com.luxoft.shop.dao.jdbc;

import com.luxoft.shop.entity.Product;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class JdbcProductDaoTest {

    @Test
    public  void testFindAllReturnData() {
        JdbcProductDao jdbcProductDao = new JdbcProductDao();
        List<Product> products = jdbcProductDao.findAll();

        assertFalse(products.isEmpty());
        for (Product product : products) {
            assertNotEquals(0, product.getId());
            assertNotNull(product.getName());
            assertNotNull(product.getPrice());
            assertNotNull(product.getNotes());
            assertNotNull(product.getCreationDate());

        }

    }


}